Create account with user relation
