+++ talelio_backend/core/__init__.py
@@ -3,7 +3,7 @@ from os import getenv
 from flask import Flask
 from flask_cors import CORS  # type: ignore
 
-from talelio_backend.data.db_tables import CREATE_ACCOUNT_TABLE
+from talelio_backend.data.db_tables import CREATE_ACCOUNT_TABLE, CREATE_USER_TABLE
 from talelio_backend.interfaces.api.accounts.account_controller import accounts_v1
 from talelio_backend.interfaces.api.articles.article_controller import articles_v1
 from talelio_backend.interfaces.api.assets.asset_controller import assets_v1
@@ -21,6 +21,7 @@ def create_db_tables() -> None:
     with connection:
         with connection.cursor() as cursor:
             cursor.execute(CREATE_ACCOUNT_TABLE)
+            cursor.execute(CREATE_USER_TABLE)
 
     connection.close()
