+++ talelio_backend/data/db_tables.py
@@ -14,5 +14,13 @@ CREATE_ACCOUNT_TABLE = (f"""
 
 CREATE_USER_TABLE = (f"""
     CREATE TABLE IF NOT EXISTS "user"
-    ()
+    (
+        id INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+        account_id INTEGER REFERENCES account (id) ON DELETE CASCADE UNIQUE,
+        created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT (NOW() AT TIME ZONE '{TIME_ZONE}'),
+        updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT (NOW() AT TIME ZONE '{TIME_ZONE}'),
+        username VARCHAR(20) UNIQUE NOT NULL,
+        location VARCHAR(50),
+        avatar_url VARCHAR(255)
+    )
     """)
