+++ talelio_backend/shared/database/db.py
@@ -22,19 +22,6 @@ connection = psycopg2.connect(f'''
     password={db_connection_values['db_password']}
     ''')
 
-CREATE_USER_TABLE = (f"""
-    CREATE TABLE IF NOT EXISTS "user"
-    (
-        id INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-        account_id INTEGER REFERENCES account (id) ON DELETE CASCADE UNIQUE,
-        created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT (NOW() AT TIME ZONE '{TIME_ZONE}'),
-        updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT (NOW() AT TIME ZONE '{TIME_ZONE}'),
-        username VARCHAR(20) UNIQUE NOT NULL,
-        location VARCHAR(50),
-        avatar_url VARCHAR(255)
-    );
-    """)
-
 CREATE_ARTICLE_TABLE = (f"""
     CREATE TABLE IF NOT EXISTS article
     (
